requirements.txt
transformers==4.26.1
hydra-core==1.3.1
torch==1.13.1
numpy==1.24.2
pandas==1.5.3
ipykernel==6.21.2
underthesea==6.1.2